Big clean up and restructure of unit tests
