+++ src/java/org/jpl7/test/JPLTestSuite.java
@@ -1,0 +1,32 @@
+package org.jpl7.test;
+
+
+import org.junit.runner.JUnitCore;
+import org.junit.runner.Result;
+import org.junit.runner.RunWith;
+import org.junit.runner.notification.Failure;
+import org.junit.runners.Suite;
+
+@RunWith(Suite.class)
+@Suite.SuiteClasses({
+        org.jpl7.test.junit.AtomTest.class,
+        org.jpl7.test.junit.DataManagement.class,
+        org.jpl7.test.junit.GetSolution.class,
+        org.jpl7.test.junit.IntegerTest.class,
+        org.jpl7.test.junit.JRef.class,
+        org.jpl7.test.junit.ListTest.class,
+        org.jpl7.test.junit.ModuleTest.class,
+        org.jpl7.test.junit.MutualRecursion.class,
+        org.jpl7.test.junit.QueryBuilder.class,
+        org.jpl7.test.junit.StringTest.class,
+        org.jpl7.test.junit.TypesTest.class,
+        org.jpl7.test.junit.Unicode.class,
+        org.jpl7.test.junit.Variables.class,
+        org.jpl7.test.junit.Tests.class,
+})
+
+public class JPLTestSuite {
+    // the class remains empty,
+    // used only as a holder for the above annotations
+    }
+
