--- src/java/org/jpl7/test/MaxObjects.java
@@ -1,4 +1,0 @@
-package org.jpl7.test;
-
-public class MaxObjects {
-}
