+++ src/java/org/jpl7/test/junit/Unicode.java
@@ -1,0 +1,135 @@
+package org.jpl7.test.junit;
+
+import org.jpl7.Atom;
+import org.jpl7.JPL;
+import org.jpl7.Query;
+import org.jpl7.Term;
+import org.jpl7.fli.Prolog;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TestRule;
+import org.junit.rules.TestWatcher;
+import org.junit.runner.Description;
+
+import static org.junit.Assert.*;
+
+
+public class Unicode {
+    public static final String startup = (System.getenv("SWIPL_BOOT_FILE") == null ? "../../src/swipl.prc"
+            : System.getenv("SWIPL_BOOT_FILE"));
+    public static final String test_jpl = (System.getenv("TEST_JPL") == null ? "test_jpl.pl"
+            : System.getenv("TEST_JPL"));
+    public static final String syntax = (System.getenv("SWIPL_SYNTAX") == null ? "modern"
+            : System.getenv("SWIPL_SYNTAX"));
+    public static final String home = (System.getenv("SWI_HOME_DIR") == null ? "../.."
+            : System.getenv("SWI_HOME_DIR"));
+
+
+
+	public static void main(String argv[]) {
+    }
+
+    ///////////////////////////////////////////////////////////////////////////////
+    // TESTING CONFIGURATION
+    ///////////////////////////////////////////////////////////////////////////////
+
+    /**
+     * This is done at the class loading, before any test is run
+     */
+    @BeforeClass
+    public static void setUp() {
+        if (syntax.equals("traditional")) {
+            JPL.setTraditional();
+            Prolog.set_default_init_args(new String[] {
+//					"libswipl.dll", "-x", startup, "-f", "none",
+                    "libswipl.dll", "-f", "none",
+                    "-g", "true", "--traditional", "-q",
+                    "--home="+home, "--no-signals", "--no-packs" });
+        } else {
+            Prolog.set_default_init_args(new String[] {
+//					"libswipl.dll", "-x", startup, "-f", "none",
+                    "libswipl.dll", "-f", "none",
+                    "-g", "true", "-q",
+                    "--home="+home, "--no-signals", "--no-packs" });
+        }
+    }
+
+    @Rule
+    public TestRule watcher = new TestWatcher() {
+        protected void starting(Description description) {
+//            logger.info("{} being run...", description.getMethodName());
+
+            System.out.println("Starting test: " + description.getMethodName());
+        }
+    };
+
+
+
+
+    ///////////////////////////////////////////////////////////////////////////////
+    // SUPPORTING CODE
+    ///////////////////////////////////////////////////////////////////////////////
+
+
+
+    ///////////////////////////////////////////////////////////////////////////////
+    // TESTS
+    ///////////////////////////////////////////////////////////////////////////////
+
+    @Test
+    public void testUnicode0() {
+        assertTrue(Query.hasSolution("atom_codes(?,[32])", new Term[]{new Atom(" ")}));
+    }
+
+    @Test
+    public void testUnicode0a() {
+        assertTrue(Query.hasSolution("atom_codes(?,[32])", new Term[]{new Atom("\u0020")}));
+    }
+
+    @Test
+    public void testUnicode0b() {
+        assertTrue(Query.hasSolution("atom_codes(?,[0])", new Term[]{new Atom("\u0000")}));
+    }
+
+    @Test
+    public void testUnicode0c() {
+        assertTrue(Query.hasSolution("atom_codes(?,[1])", new Term[]{new Atom("\u0001")}));
+    }
+
+    @Test
+    public void testUnicode0d() {
+        assertTrue(Query.hasSolution("atom_codes(?,[127])", new Term[]{new Atom("\u007F")}));
+    }
+
+    @Test
+    public void testUnicode0e() {
+        assertTrue(Query.hasSolution("atom_codes(?,[128])", new Term[]{new Atom("\u0080")}));
+    }
+
+    @Test
+    public void testUnicode0f() {
+        assertTrue(Query.hasSolution("atom_codes(?,[255])", new Term[]{new Atom("\u00FF")}));
+    }
+
+    @Test
+    public void testUnicode0g() {
+        assertTrue(Query.hasSolution("atom_codes(?,[256])", new Term[]{new Atom("\u0100")}));
+    }
+
+    @Test
+    public void testUnicode1() {
+        assertTrue(Query.hasSolution("atom_codes(?,[0,127,128,255])",
+                new Term[]{new Atom("\u0000\u007F\u0080\u00FF")}));
+    }
+
+    @Test
+    public void testUnicode2() {
+        assertTrue(Query.hasSolution("atom_codes(?,[256,32767,32768,65535])",
+                new Term[]{new Atom("\u0100\u7FFF\u8000\uFFFF")}));
+    }
+
+
+
+}
