+++ src/java/org/jpl7/test/standalone/Family.java
@@ -1,4 +1,4 @@
-package org.jpl7.test;
+package org.jpl7.test.standalone;
 
 import java.util.Map;
